Fix deployment errors, project structure and directories used for commands
