+++ Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
 members = [
-    "src/icp_rust_boilerplate_backend",
+    "src/social_media_app",
 ]
