+++ dfx.json
@@ -1,10 +1,10 @@
 {
   "version": 1,
   "canisters": {
-    "icp_rust_boilerplate_backend": {
+    "social_media_app": {
       "type": "rust",
-      "package": "icp_rust_boilerplate_backend",
-      "candid": "src/icp_rust_boilerplate_backend/icp_rust_boilerplate_backend.did"
+      "package": "social_media_app",
+      "candid": "src/social_media_app/social_media_app.did"
     }
   },
   "output_env_file": ".env"
