+++ did.sh
@@ -12,7 +12,7 @@ function generate_did() {
   candid-extractor "target/wasm32-unknown-unknown/release/$canister.wasm" > "$canister_root/$canister.did"
 }
 
-CANISTERS=icp_rust_boilerplate_backend
+CANISTERS=social_media_app
 
 for canister in $(echo $CANISTERS | sed "s/,/ /g")
 do
