--- icp_rust_boilerplate_backend.d
@@ -1,1 +1,0 @@
-/Users/mac/icp\ social\ dapp/target/wasm32-unknown-unknown/release/icp_rust_boilerplate_backend.wasm: /Users/mac/icp\ social\ dapp/src/icp_rust_boilerplate_backend/src/lib.rs
