+++ src/social_media_app/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "social_media_app"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+candid = "0.9.9"
+ic-cdk = "0.11.1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0"
+ic-stable-structures = "0.5.6"
